backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.60.0
sse-starlette>=2.2.0
python-dotenv>=1.0.0
